Fix mouse boundary jail
In some cases when panning with the mouse very
near the edge of the screen, or when the mouse
moves too quickly, issuing a right-click action
can open a Windows right-click context menu and
trap the mouse outside of the client window. See
attached video for the behavior before and after.

This isn't perfect and there are still some issues
with the sprite trailing the mouse cursor, but
this seems to at least fix the most dangerous part
of this bug (for me).